Improve the heuristic strategy of choosing one deb pkg from many
Previously, we chose the deb package whose name has the smallest
lexical order.  This worked for some cases but did not work well when
there are also `lib32` and `libx32` packages to choose.

Now, we choose the deb package whose name has the smallest edit
distance with the pname of the nix package providing the soname.  This
simple heuristic strategy seems to work pretty well.