Deb: ensure a nix version is valid as a deb version
Below is the info about nix and deb version formats.
- deb version specification[1]
  - summary: contain `A-Za-z0-9`, `.`, `-`, `+` and `~`
- nix has no version specification
  - summary: contain anything
  - Nixpkgs doc[2]: starts with number
  - `builtins.compareVersions` implementation[3]
    - contain anything
    - use `.` and `-` as separator
  - `nix-env --upgrade` doc[4]

Technically, a nix version format is a suprset of a deb version
format.  However, I do not think it is common that a nix version is
invalid as a deb version.  As a result, we currently exit if that
happens.  We can improve the logic to handle invalid nix versions in
the future if those cases are not rare.

[1]: https://www.debian.org/doc/debian-policy/ch-controlfields.html#version
[2]: https://github.com/NixOS/nixpkgs/blob/19059aa9c38d64f62f7e9828d4d0277debbacc19/pkgs/README.md?plain=1#L402-L421
[3]: https://github.com/NixOS/nix/blob/74a107d3643ef7deaea42d2708556a204384d503/src/libstore/names.cc#L59-L110
[4]: https://nix.dev/manual/nix/2.29/command-ref/nix-env/upgrade.html#versions